Remove Remote Buckets section
