+++ portal-ui/src/screens/Console/Console.tsx
@@ -67,7 +67,6 @@ import { ISessionResponse } from "./types";
 import { saveSessionResponse } from "./actions";
 import TenantDetails from "./Tenants/TenantDetails/TenantDetails";
 import { clearSession } from "../../common/utils";
-import RemoteBuckets from "./RemoteBuckets/RemoteBuckets";
 import ObjectBrowser from "./ObjectBrowser/ObjectBrowser";
 import ListObjects from "./Buckets/ListBuckets/Objects/ListObjects/ListObjects";
 
@@ -278,10 +277,6 @@ const Console = ({
     {
       component: Policies,
       path: "/policies",
-    },
-    {
-      component: RemoteBuckets,
-      path: "/remote-buckets",
     },
     {
       component: Trace,
