+++ portal-ui/src/screens/Console/Menu/Menu.tsx
@@ -21,7 +21,6 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import WebAssetIcon from "@material-ui/icons/WebAsset";
 import HealingIcon from "@material-ui/icons/Healing";
-import CloudUploadIcon from "@material-ui/icons/CloudUpload";
 import DescriptionIcon from "@material-ui/icons/Description";
 import FileCopyIcon from "@material-ui/icons/FileCopy";
 import Collapse from "@material-ui/core/Collapse";
@@ -246,14 +245,6 @@ const Menu = ({ userLoggedIn, classes, pages }: IMenuProps) => {
       to: "/policies",
       name: "IAM Policies",
       icon: <IAMPoliciesIcon />,
-    },
-    {
-      group: "Admin",
-      type: "item",
-      component: NavLink,
-      to: "/remote-buckets",
-      name: "Remote Buckets",
-      icon: <CloudUploadIcon />,
     },
     {
       group: "Tools",
