--- portal-ui/src/screens/Console/RemoteBuckets/types.ts
@@ -1,32 +1,0 @@
-// This file is part of MinIO Console Server
-// Copyright (c) 2020 MinIO, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-export interface IRemoteBucketsResponse {
-  buckets: IRemoteBucket[];
-  total: number;
-}
-
-export interface IRemoteBucket {
-  name: string;
-  accessKey: string;
-  secretKey: string;
-  sourceBucket: string;
-  targetURL: string;
-  targetBucket: string;
-  remoteARN: string;
-  status: string;
-  service: string;
-}
